Week-9/lambda_requirements.txt
numpy==1.26.1
pill
